Renamed engine, added shader compilation step to build, started initializing graphics pipeline
