+++ Engine/CPPLINT.cfg
@@ -1,0 +1,4 @@
+set noparent
+
+filter=-legal
+linelength=120
