+++ Engine/lib_src/Core/Core.h
@@ -1,0 +1,7 @@
+#pragma once
+#include "Logger/Logger.h"
+#include <vector>
+
+namespace Core {
+	std::vector<char> readFileBinary(const std::string& filename);
+};
