+++ Engine/lib_src/Testing.cpp
@@ -1,13 +1,13 @@
 #include "Testing.h"
 #define GLFW_INCLUDE_VULKAN
 #include <GLFW/glfw3.h>
+#include <iostream>
 
 #define GLM_FORCE_RADIANS
 #define GLM_FORCE_DEPTH_ZERO_TO_ONE
 #include <glm/vec4.hpp>
 #include <glm/mat4x4.hpp>
 
-#include <iostream>
 #include <boost/random.hpp>
 
 int test() {
@@ -21,7 +21,7 @@ int test() {
     glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
     GLFWwindow* window = glfwCreateWindow(800, 600, "Vulkan window", nullptr, nullptr);
 
-  uint32_t extensionCount = 0;
+    uint32_t extensionCount = 0;
 
     vkEnumerateInstanceExtensionProperties(nullptr, &extensionCount, nullptr);
