+++ Engine/lib_src/Testing.h
@@ -1,0 +1,3 @@
+#pragma once
+#include "GraphicsLib/VulkanInit.h"
+int test();
