+++ Engine/src/main.cpp
@@ -1,4 +1,4 @@
-#include <iostream>
+
 #include "main.h"
 #include "Testing.h"
