--- PenixEngine/CPPLINT.cfg
@@ -1,3 +1,0 @@
-set noparent
-
-filter=-legal
