--- PenixEngine/lib_src/Core/Core.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "Logger/Logger.h"
