--- PenixEngine/lib_src/PenixEngine.cpp
@@ -1,10 +1,0 @@
-﻿// PenixEngine.cpp: определяет точку входа для приложения.
-//
-
-#include "PenixEngine.h"
-using namespace std;
-#include "Testing.h"
-
-void fn() {
-
-};
