--- PenixEngine/lib_src/PenixEngine.h
@@ -1,9 +1,0 @@
-﻿// PenixEngine.h : включаемый файл для стандартных системных включаемых файлов
-// или включаемые файлы для конкретного проекта.
-
-#pragma once
-
-#include <iostream>
-
-// TODO: установите здесь ссылки на дополнительные заголовки, требующиеся для программы.
-void fn();
