--- PenixEngine/lib_src/Testing.h
@@ -1,3 +1,0 @@
-#pragma once
-#include "GraphicsLib/VulkanTest.h"
-int test();
